move fake bundles spec into conftest to avoid collision
